Removed deprecated Logger and LoggerFactory
Also, removed unused "fatal" and templated logging methods from LogDelegate SPI.
