+++ src/main/java/io/vertx/core/logging/JULLogDelegate.java
@@ -45,102 +45,44 @@ public class JULLogDelegate implements LogDelegate {
     return logger.isLoggable(Level.FINEST);
   }
 
-  public void fatal(final Object message) {
-    log(Level.SEVERE, message);
-  }
-
-  public void fatal(final Object message, final Throwable t) {
-    log(Level.SEVERE, message, t);
-  }
-
   public void error(final Object message) {
     log(Level.SEVERE, message);
-  }
-
-  @Override
-  public void error(Object message, Object... params) {
-    log(Level.SEVERE, message, null, params);
   }
 
   public void error(final Object message, final Throwable t) {
     log(Level.SEVERE, message, t);
   }
 
-  @Override
-  public void error(Object message, Throwable t, Object... params) {
-    log(Level.SEVERE, message, t, params);
-  }
-
   public void warn(final Object message) {
     log(Level.WARNING, message);
-  }
-
-  @Override
-  public void warn(Object message, Object... params) {
-    log(Level.WARNING, message, null, params);
   }
 
   public void warn(final Object message, final Throwable t) {
     log(Level.WARNING, message, t);
   }
 
-  @Override
-  public void warn(Object message, Throwable t, Object... params) {
-    log(Level.WARNING, message, t, params);
-  }
-
   public void info(final Object message) {
     log(Level.INFO, message);
-  }
-
-  @Override
-  public void info(Object message, Object... params) {
-    log(Level.INFO, message, null, params);
   }
 
   public void info(final Object message, final Throwable t) {
     log(Level.INFO, message, t);
   }
 
-  @Override
-  public void info(Object message, Throwable t, Object... params) {
-    log(Level.INFO, message, t, params);
-  }
-
   public void debug(final Object message) {
     log(Level.FINE, message);
-  }
-
-  @Override
-  public void debug(Object message, Object... params) {
-    log(Level.FINE, message, null, params);
   }
 
   public void debug(final Object message, final Throwable t) {
     log(Level.FINE, message, t);
   }
 
-  @Override
-  public void debug(Object message, Throwable t, Object... params) {
-    log(Level.FINE, message, t, params);
-  }
-
   public void trace(final Object message) {
     log(Level.FINEST, message);
   }
 
-  @Override
-  public void trace(Object message, Object... params) {
-    log(Level.FINEST, message, null, params);
-  }
-
   public void trace(final Object message, final Throwable t) {
     log(Level.FINEST, message, t);
-  }
-
-  @Override
-  public void trace(Object message, Throwable t, Object... params) {
-    log(Level.FINEST, message, t, params);
   }
 
   private void log(Level level, Object message) {
