+++ src/main/java/io/vertx/core/logging/Log4j2LogDelegate.java
@@ -11,6 +11,7 @@
 
 package io.vertx.core.logging;
 
+import io.vertx.core.impl.logging.Logger;
 import io.vertx.core.spi.logging.LogDelegate;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.message.FormattedMessage;
@@ -26,7 +27,6 @@ public class Log4j2LogDelegate implements LogDelegate {
 
   final ExtendedLogger logger;
 
-  @SuppressWarnings("deprecation")
   final static String FQCN = Logger.class.getCanonicalName();
 
   Log4j2LogDelegate(final String name) {
@@ -50,102 +50,44 @@ public class Log4j2LogDelegate implements LogDelegate {
     return logger.isTraceEnabled();
   }
 
-  public void fatal(final Object message) {
-    log(Level.FATAL, message);
-  }
-
-  public void fatal(final Object message, final Throwable t) {
-    log(Level.FATAL, message, t);
-  }
-
   public void error(final Object message) {
     log(Level.ERROR, message);
-  }
-
-  @Override
-  public void error(Object message, Object... params) {
-    log(Level.ERROR, message.toString(), params);
   }
 
   public void error(final Object message, final Throwable t) {
     log(Level.ERROR, message, t);
   }
 
-  @Override
-  public void error(Object message, Throwable t, Object... params) {
-    log(Level.ERROR, message.toString(), t, params);
-  }
-
   public void warn(final Object message) {
     log(Level.WARN, message);
-  }
-
-  @Override
-  public void warn(Object message, Object... params) {
-    log(Level.WARN, message.toString(), params);
   }
 
   public void warn(final Object message, final Throwable t) {
     log(Level.WARN, message, t);
   }
 
-  @Override
-  public void warn(Object message, Throwable t, Object... params) {
-    log(Level.WARN, message.toString(), t, params);
-  }
-
   public void info(final Object message) {
     log(Level.INFO, message);
-  }
-
-  @Override
-  public void info(Object message, Object... params) {
-    log(Level.INFO, message.toString(), params);
   }
 
   public void info(final Object message, final Throwable t) {
     log(Level.INFO, message, t);
   }
 
-  @Override
-  public void info(Object message, Throwable t, Object... params) {
-    log(Level.INFO, message.toString(), t, params);
-  }
-
   public void debug(final Object message) {
     log(Level.DEBUG, message);
-  }
-
-  @Override
-  public void debug(Object message, Object... params) {
-    log(Level.DEBUG, message.toString(), params);
   }
 
   public void debug(final Object message, final Throwable t) {
     log(Level.DEBUG, message, t);
   }
 
-  @Override
-  public void debug(Object message, Throwable t, Object... params) {
-    log(Level.DEBUG, message.toString(), t, params);
-  }
-
   public void trace(final Object message) {
     log(Level.TRACE, message);
   }
 
-  @Override
-  public void trace(Object message, Object... params) {
-    log(Level.TRACE, message.toString(), params);
-  }
-
   public void trace(final Object message, final Throwable t) {
     log(Level.TRACE, message.toString(), t);
-  }
-
-  @Override
-  public void trace(Object message, Throwable t, Object... params) {
-    log(Level.TRACE, message.toString(), t, params);
   }
 
   private void log(Level level, Object message) {
