+++ src/main/java/io/vertx/core/logging/SLF4JLogDelegate.java
@@ -28,8 +28,7 @@ public class SLF4JLogDelegate implements LogDelegate {
 
   private static final Object[] EMPTY_PARAMETERS = {};
 
-  @SuppressWarnings("deprecation")
-  private static final String FQCN = io.vertx.core.logging.Logger.class.getCanonicalName();
+  private static final String FQCN = Logger.class.getCanonicalName();
 
   private final Logger logger;
 
@@ -58,100 +57,44 @@ public class SLF4JLogDelegate implements LogDelegate {
     return logger.isTraceEnabled();
   }
 
-  public void fatal(final Object message) {
-    log(ERROR_INT, message);
-  }
-
-  public void fatal(final Object message, final Throwable t) {
-    log(ERROR_INT, message, t);
-  }
-
   public void error(final Object message) {
     log(ERROR_INT, message);
-  }
-
-  @Override
-  public void error(Object message, Object... params) {
-    log(ERROR_INT, message, null, params);
   }
 
   public void error(final Object message, final Throwable t) {
     log(ERROR_INT, message, t);
   }
 
-  @Override
-  public void error(Object message, Throwable t, Object... params) {
-    log(ERROR_INT, message, t, params);
-  }
-
   public void warn(final Object message) {
     log(WARN_INT, message);
-  }
-
-  @Override
-  public void warn(Object message, Object... params) {
-    log(WARN_INT, message, null, params);
   }
 
   public void warn(final Object message, final Throwable t) {
     log(WARN_INT, message, t);
   }
 
-  @Override
-  public void warn(Object message, Throwable t, Object... params) {
-    log(WARN_INT, message, t, params);
-  }
-
   public void info(final Object message) {
     log(INFO_INT, message);
-  }
-
-  @Override
-  public void info(Object message, Object... params) {
-    log(INFO_INT, message, null, params);
   }
 
   public void info(final Object message, final Throwable t) {
     log(INFO_INT, message, t);
   }
 
-  @Override
-  public void info(Object message, Throwable t, Object... params) {
-    log(INFO_INT, message, t, params);
-  }
-
   public void debug(final Object message) {
     log(DEBUG_INT, message);
-  }
-
-  public void debug(final Object message, final Object... params) {
-    log(DEBUG_INT, message, null, params);
   }
 
   public void debug(final Object message, final Throwable t) {
     log(DEBUG_INT, message, t);
   }
 
-  public void debug(final Object message, final Throwable t, final Object... params) {
-    log(DEBUG_INT, message, t, params);
-  }
-
   public void trace(final Object message) {
     log(TRACE_INT, message);
   }
 
-  @Override
-  public void trace(Object message, Object... params) {
-    log(TRACE_INT, message, null, params);
-  }
-
   public void trace(final Object message, final Throwable t) {
     log(TRACE_INT, message, t);
-  }
-
-  @Override
-  public void trace(Object message, Throwable t, Object... params) {
-    log(TRACE_INT, message, t, params);
   }
 
   private void log(int level, Object message) {
