+++ src/test/java/io/vertx/core/impl/logging/NoExceptionInInitializerErrorTest.java
@@ -20,8 +20,7 @@ import org.apache.logging.log4j.core.pattern.PatternConverter;
 import org.junit.After;
 import org.junit.Test;
 
-import static io.vertx.core.logging.LoggerFactory.LOGGER_DELEGATE_FACTORY_CLASS_NAME;
-
+import static io.vertx.core.impl.logging.LoggerFactory.LOGGER_DELEGATE_FACTORY_CLASS_NAME;
 
 /**
  * Must be separated from {@link LoggingBackendSelectionTest}.
