+++ src/test/java/io/vertx/core/logging/LoggerFactoryTest.java
@@ -11,6 +11,7 @@
 
 package io.vertx.core.logging;
 
+import io.vertx.core.impl.logging.LoggerFactory;
 import org.junit.Test;
 
 /**
