+++ src/test/java/io/vertx/it/SLF4JLogDelegateTest.java
@@ -11,8 +11,9 @@
 
 package io.vertx.it;
 
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
+import io.vertx.core.impl.logging.Logger;
+import io.vertx.core.impl.logging.LoggerAdapter;
+import io.vertx.core.impl.logging.LoggerFactory;
 import io.vertx.core.logging.SLF4JLogDelegate;
 import io.vertx.core.spi.logging.LogDelegate;
 import org.junit.Test;
@@ -28,19 +29,18 @@ import java.util.concurrent.Callable;
 import static org.junit.Assert.*;
 
 /**
- * Theses test checks the SLF4J log delegate. It assumes the binding used by SLF4J is slf4j-simple with the default
+ * These test checks the SLF4J log delegate. It assumes the binding used by SLF4J is slf4j-simple with the default
  * configuration. It injects a print stream to read the logged message. This is definitely a hack, but it's the only
  * way to test the output.
  *
  * @author <a href="http://escoffier.me">Clement Escoffier</a>
  */
-@SuppressWarnings("deprecation")
 public class SLF4JLogDelegateTest {
 
   @Test
   public void testDelegateUnwrap() {
     Logger logger = LoggerFactory.getLogger("my-slf4j-logger");
-    LogDelegate delegate = logger.getDelegate();
+    LogDelegate delegate = ((LoggerAdapter) logger).unwrap();
     assertNotNull("Delegate is null", delegate);
     try {
       org.slf4j.Logger unwrapped = (org.slf4j.Logger) delegate.unwrap();
@@ -57,7 +57,7 @@ public class SLF4JLogDelegateTest {
 
   @Test
   public void testInfoLocationAware() {
-    testInfo(new Logger(new SLF4JLogDelegate(new TestLocationAwareLogger("my-slf4j-logger"))));
+    testInfo(new LoggerAdapter(new SLF4JLogDelegate(new TestLocationAwareLogger("my-slf4j-logger"))));
   }
 
   private void testInfo(Logger logger) {
@@ -67,26 +67,6 @@ public class SLF4JLogDelegateTest {
     result = record(() -> logger.info("exception", new NullPointerException()));
     assertTrue(result.contains("[main] INFO my-slf4j-logger - exception"));
     assertTrue(result.contains("java.lang.NullPointerException"));
-
-    result = record(() -> logger.info("hello {} and {}", "Paulo", "Julien"));
-    assertContains("[main] INFO my-slf4j-logger - hello Paulo and Julien", result);
-
-    result = record(() -> logger.info("hello {}", "vert.x"));
-    assertContains("[main] INFO my-slf4j-logger - hello vert.x", result);
-
-    result = record(() -> logger.info("hello {} - {}", "vert.x"));
-    assertContains("[main] INFO my-slf4j-logger - hello vert.x - {}", result);
-
-    result = record(() -> logger.info("hello {}", "vert.x", "foo"));
-    assertContains("[main] INFO my-slf4j-logger - hello vert.x", result);
-
-    result = record(() -> logger.info("{}, an exception has been thrown", new IllegalStateException(), "Luke"));
-    assertTrue(result.contains("[main] INFO my-slf4j-logger - Luke, an exception has been thrown"));
-    assertTrue(result.contains("java.lang.IllegalStateException"));
-
-    result = record(() -> logger.info("{}, an exception has been thrown", "Luke", new IllegalStateException()));
-    assertTrue(result.contains("[main] INFO my-slf4j-logger - Luke, an exception has been thrown"));
-    assertTrue(result.contains("java.lang.IllegalStateException"));
   }
 
   @Test
@@ -106,26 +86,6 @@ public class SLF4JLogDelegateTest {
     result = record(() -> logger.error("exception", new NullPointerException()));
     assertTrue(result.contains("[main] ERROR my-slf4j-logger - exception"));
     assertTrue(result.contains("java.lang.NullPointerException"));
-
-    result = record(() -> logger.error("hello {} and {}", "Paulo", "Julien"));
-    assertContains("[main] ERROR my-slf4j-logger - hello Paulo and Julien", result);
-
-    result = record(() -> logger.error("hello {}", "vert.x"));
-    assertContains("[main] ERROR my-slf4j-logger - hello vert.x", result);
-
-    result = record(() -> logger.error("hello {} - {}", "vert.x"));
-    assertContains("[main] ERROR my-slf4j-logger - hello vert.x - {}", result);
-
-    result = record(() -> logger.error("hello {}", "vert.x", "foo"));
-    assertContains("[main] ERROR my-slf4j-logger - hello vert.x", result);
-
-    result = record(() -> logger.error("{}, an exception has been thrown", new IllegalStateException(), "Luke"));
-    assertTrue(result.contains("[main] ERROR my-slf4j-logger - Luke, an exception has been thrown"));
-    assertTrue(result.contains("java.lang.IllegalStateException"));
-
-    result = record(() -> logger.error("{}, an exception has been thrown", "Luke", new IllegalStateException()));
-    assertTrue(result.contains("[main] ERROR my-slf4j-logger - Luke, an exception has been thrown"));
-    assertTrue(result.contains("java.lang.IllegalStateException"));
   }
 
   private void assertContains(String expectedExcerpt, String object) {
@@ -150,26 +110,6 @@ public class SLF4JLogDelegateTest {
     result = record(() -> logger.warn("exception", new NullPointerException()));
     assertTrue(result.contains("[main] WARN my-slf4j-logger - exception"));
     assertTrue(result.contains("java.lang.NullPointerException"));
-
-    result = record(() -> logger.warn("hello {} and {}", "Paulo", "Julien"));
-    assertContains("[main] WARN my-slf4j-logger - hello Paulo and Julien", result);
-
-    result = record(() -> logger.warn("hello {}", "vert.x"));
-    assertContains("[main] WARN my-slf4j-logger - hello vert.x", result);
-
-    result = record(() -> logger.warn("hello {} - {}", "vert.x"));
-    assertContains("[main] WARN my-slf4j-logger - hello vert.x - {}", result);
-
-    result = record(() -> logger.warn("hello {}", "vert.x", "foo"));
-    assertContains("[main] WARN my-slf4j-logger - hello vert.x", result);
-
-    result = record(() -> logger.warn("{}, an exception has been thrown", new IllegalStateException(), "Luke"));
-    assertTrue(result.contains("[main] WARN my-slf4j-logger - Luke, an exception has been thrown"));
-    assertTrue(result.contains("java.lang.IllegalStateException"));
-
-    result = record(() -> logger.warn("{}, an exception has been thrown", "Luke", new IllegalStateException()));
-    assertTrue(result.contains("[main] WARN my-slf4j-logger - Luke, an exception has been thrown"));
-    assertTrue(result.contains("java.lang.IllegalStateException"));
   }
 
   private void setStream(PrintStream stream) {
@@ -210,7 +150,7 @@ public class SLF4JLogDelegateTest {
 
     TestLocationAwareLogger(String name) {
       Logger logger = LoggerFactory.getLogger(name);
-      SLF4JLogDelegate delegate = (SLF4JLogDelegate) logger.getDelegate();
+      SLF4JLogDelegate delegate = (SLF4JLogDelegate) ((LoggerAdapter) logger).unwrap().unwrap();
       this.actual = (org.slf4j.Logger) delegate.unwrap();
     }
